Changed project structure
+added config reading layer
