--- docs/utilities/ss-procfs.md
@@ -1,17 +1,0 @@
-**Сравнительный анализ утилиты ss и procfs.**
-
-Для того, чтобы понять к каким файлам обращается утилита ss, воспользуемся strace:
-
-![image](https://github.com/Saeshnikov/Linux-monitoring-utility/assets/117933964/266fd3bd-829f-41e4-8516-b1b6225edcd7)
-
-
-Файлы proc/, которые использует ss:
-
-| Файл | Описание |
-| --- | --- |
-| proc/filesystems | Этот файл отображает список типов файловых систем, поддерживаемых в настоящее время ядром. Первый столбец указывает, установлена ли файловая система на блочном устройстве. Файлы, начинающиеся с nodev, не монтируются на устройстве. Во втором столбце перечислены имена поддерживаемых файловых систем. |
-| proc/sys/net/ipv4/ip\_local\_port\_range | Этот файл указывает диапазон портов, которые будут использоваться TCP или UDP, когда требуется локальный порт. Первое число — это самый низкий используемый порт, а второе число указывает самый высокий порт. |
-
-Вывод.
-
-Из информации, которую актуально собирать в рамках нашего проекта, утилита ss обращается к /proc только для сбора локального диапазона портов. А вся остальная рассматриваемая сетевая статистика собирается из других источников. Следовательно, заменить утилиту ss на procfs в данном случае не получится.
