--- go.mod
@@ -1,3 +1,0 @@
-module abcd
-
-go 1.22.0
