+++ prototype/docs/components/hardware.md
