+++ prototype/docs/report/network.md
@@ -1,0 +1,8 @@
+| **Пункт мониторинга** | **Утилита для получения информации** | **Актуальность** | **Обоснование выбора утилиты** |
+| --- | --- | --- | --- |
+| Тип подключения | ss (txt) / ss \> file (file) <br>Вывод:<br>полный список активных соединений (TCP и сокеты). <br>Поле Netid — тип подключения:- u\_dgr — тип сокета, который работает без установки соединения;<br>- u\_str — тип сокета, который работает с предварительной установкой соединения;<br>- tcp — tcp соединение;<br>- udp — udp соединение.<br><br>ss -t, --tcp: отображать TCP-сокеты.<br>ss -u, --udp: UDP-сокеты.<br>ss -d, --dccp: DCCP сокеты.<br>ss -w, --raw: RAW сокеты.<br>ss -x, --unix: доменные сокеты Unix (псевдоним для -f unix).<br>ss -S, --sctp: SCTP-сокеты. | Необходимо понимать, как конкретно общаются два процесса| ss удобно пользоваться в данном контексте, так как она собирает обширную информацию по всем видам сокетов, а также есть множество ключей, чтобы вычленять более детальную информацию |
+| Состояния соединения | ss (txt) / ss \> file (file)<br>Вывод:<br>Поле State — состояние соединения (сеанса TCP):CLOSED — начальное состояние; LISTEN — сервер ожидает запросы от клиента; SYN-SENT — клиент хочет установить соединение с сервером и ожидает подтверждение; SYN-RECEIVED — сервер получил запрос на создание сеанса, отправил ответный запрос и ожидает подтверждение; ESTABLISHED — соединение установлено, идет передача данных; FIN-WAIT-1 — одна из сторон завершает соединение, отправив флаг FIN; CLOSE-WAIT — другая сторона переходит в это состояние, отправив подтверждение на FIN, но продолжает передачу; FIN-WAIT-2 — первый узел получил ACK, разорвал свое соединение, но еще читает данные; LAST-ACK — второй узел заканчивает передачу и отправляет флаг FIN; TIME-WAIT — сервер получил пакет с флагом FIN, отправил флаг ACK и ждёт некоторое время, перед окончательным закрытием соединения; CLOSING — обе стороны инициировали закрытие соединения одновременно. | Понимание состояния соединения позволит дать более полное понимание о возникающих ошибках подключения и потери пакетов||
+| Процессы, использующие сокеты | ss -p (txt) / ss -p \> file (file) | Зная, какому процессу принадлежит сокет, можно обратиться к трассировке системных вызовов и посмотреть, в чем проблема, если такова возникает ||
+| Домен соединения; Семантика коммуникации, Путь по которому сокет связан с дескриптором;Связь между двумя процессами по сети| strace -e trace=socket,connect,send,recv -o {file} {process}<br><br>- socket(AF\_UNIX, SOCK\_STREAM, 0) =sockfd ; параметры: Домен соединения (локальное соединение, IPv4 протоколы Интернет, IPv6 протоколы Интернет, IPX, устройство для взаимодействия с ядром и тд)Семантика коммуникации(SOCK\_STREAM двусторонний надежный и последовательный поток байтов, SOCK\_DGRAM поддерживает датаграммы, и тд)<br><br>-connect() - устанавливает соединение с конечной точкой внешней сети | Понимание того, что процесс работает с чем-то через сокет по определенному пути. Таким образом через, например, утилиту ss, можно найти процесс, привязанный к этому сокету с другой стороны.| Эта утилита позволяет установить использует ли процесс данное компьютерное соединение для передачи данных и/или получения и что передается |
+| Информация о файлах, используемых Unix сокетами | lsof -U<br>Вывод:<br>Имя команды, PID, имя пользователя, файловый дескриптор | Может быть полезно в случае более детального анализа общения процессов по Unix сокетам | lsof удобно использовать для просмотра и предоставления информации о файлах, используемых процессами |
+| Сетевой трафик конкретного типа | tcpdump icmp/ tcp, udp<br>Вывод:<br>Метка времени Unix, протокол, имя или IP-адрес исходного хоста и номер порта, имя хоста или IP-адрес назначения и номер порта, флаги TCP, порядковый номер данных в пакете, номер подтверждения, размер окна (количество байт, доступных в принимающем буфере), количество байтов, доступных в приемном буфере., параметры TCP, длина полезной нагрузки данных. | ??? | Утилита tcpdump - удобный инструмент для решения проблем сетевого подключения. Несмотря на огромный вывод, который выдает утилита, ей удобно пользоваться благодаря множеству фильтров и выражений, которые можно комбинировать. |
