+++ prototype/docs/utilities/lsof-procfs.md
@@ -1,0 +1,16 @@
+**Сравнительный анализ утилиты**  **lsof**  **и procfs.**
+
+Для того, чтобы понять к каким файлам обращается утилита lsof, воспользуемся strace:
+
+![image](https://github.com/Saeshnikov/Linux-monitoring-utility/assets/117933964/8a1b461c-71a0-47d1-979d-e10a1127d907)
+
+
+Утилита lsof обращается ко многим файлам /proc, но главное, что нас интересует в данном проекте – это файл /proc/fdinfo.
+
+| Файлы | Описание |
+| --- | --- |
+| /proc/\<pid\>/fdinfo/\<fdinfo\> | Этот целевой объект предоставляет файлы (\<fdinfo\>), имена которых являются десятичными представлениями дескрипторов открытых файлов в пределах указанного идентификатора процесса (\<pid\>). Файл может быть прочитан процессом-владельцем для получения информации о сеансе открытия процесса с файлом или каталогом.<br/>Эта информация состоит из одной или нескольких пар имя-значение. Пробелы отделяют имена от значений, а новые строки отделяют пары имя-значение таким образом, чтобы в каждой строке была одна пара имя-значение. Файл содержит следующие поля:<br/>- pos - показывает текущее положение или смещение в рамках открытого сеанса.<br/>- флаги - отображает флаги дескриптора открытого файла.<br/>Также может существовать другая информация. Например, поле сокета включено для отображения значения дескриптора сокета, если открытый сеанс связан со специальным файлом типа сокета. |
+
+Вывод.
+
+Информацию о том, какие файлы используются теми или иными процессами, утилита lsof собирает из файлов /proc. Однако, lsof в себе сортирует и компонует всю эту собранную информацию, что делает её вывод более полезным, чем «сырая» информация procfs.
