+++ prototype/docs/utilities/tcpdump.md
@@ -1,0 +1,168 @@
+**Утилита tcpdump.**
+
+Утилита tcpdump — инструмент командной строки, который способен перехватывать и анализировать сетевой трафик. tcpdump позволяет захватывать не только TCP-трафик, а и UDP, ARP или ICMP. Захваченные пакеты могут быть записаны в файл или стандартный вывод.
+
+**Основы tcpdump.**
+
+Запускать tcpdump нужно с правами root. Если запускать утилиту без дополнительных опций, произойдет перехват всех пакетов, проходящих через интерфейс по умолчанию. tcpdump позволяет проверять заголовки пакетов TCP/IP и выводить одну строку для каждого из пакетов. Она будет делать это до тех пор, пока не нажать Ctrl + C.
+
+Вывод:
+
+Каждая строка включает:
+
+- Метка времени Unix ( **20: 58: 26.765637** )
+- Протокол ( **IP** )
+- Имя или IP-адрес исходного хоста и номер порта ( **10.0.0.50.80** )
+- Имя хоста или IP-адрес назначения и номер порта ( **10.0.0.1.53181** )
+- Флаги TCP (**Flags [F.]**). Указывают на состояние соединения и могут содержать более одного значения:
+  - S — SYN. Первый шаг в установлении соединения
+  - F — FIN. Прекращение соединения
+  - . — ACK. Пакет подтверждения принят успешно
+  - P — PUSH. Указывает получателю обрабатывать пакеты вместо их буферизации
+  - R — RST. Связь прервалась
+- Порядковый номер данных в пакете. ( **seq 1** )
+- Номер подтверждения. ( **ack 2** )
+- Размер окна (количество байт, доступных в принимающем буфере) (**win 453)**.
+- Количество байтов, доступных в приемном буфере.
+- Параметры TCP.
+- Длина полезной нагрузки данных. ( **length 0** )
+
+
+**Опции.**
+
+Общий синтаксис команды tcpdump выглядит следующим образом:
+
+tcpdump [options] [expression]
+
+- Команда options позволяет вам контролировать поведение команды.
+- Фильтр expression определяет, какие пакеты будут захвачены.
+
+Базовые опции:
+
+- **-D** : вывести исчерпывающий перечень доступных интерфейсов;
+- **-i eth0** : проверить определенный интерфейс (eth0);
+- **-i any** : проверить интерфейсы на предмет наличия любого трафика;
+- **-n** : выводить IP- адреса, а не названия хостов;
+- **-nn** : выводить IP-адреса вместе с номерами портов, а не названия хостов с названиями протоколов;
+- **-q** : отображать минимальный объем данных о пакете;
+- **-t** : отключить вывод метки времени для всех строк;
+- **-tttt** : включает для каждой из строк отображение временных меток в формате по умолчанию;
+- **-X** : отображать данные пакета одновременно в шестнадцатеричной и в ASCII-кодировках;
+- **-A**: отображать данные пакета в ASCII-кодировке;
+- **-XX** : аналог -X, который также выводит ethernet header;
+- **-v, -vv, -vvv** : увеличить объем возвращаемых данных о пакетах;
+- **-c** : вывести указанное число пакетов, после чего стоп;
+- **-s** : вычислить длину snaplength захвата в байтах (-s0, чтобы вычислить все, если Вы намеренно не захватили меньше);
+- **-S** : вывести абсолютные порядковые номера;
+- **-e** : продемонстрировать ethernet header;
+- **-q** : вывести минимум данных о пакете;
+- **-E** : произвести расшифровку трафика IPSEC, отдав ключ шифрования.
+
+Утилита позволяет комбинировать опции всеми доступными способами: and либо &, or либо ||, not либо !.
+
+Утилита tcpdump также предусматривает использование спецсимволов для группирования нескольких выражений и логических операторов: поставив перед каждой из скобок обратную косую черту — \; заключив все выражение в одинарные кавычки — ' (например - tcpdump src 10.0.2.4 and \(dst port 3389 or 22\)).
+
+
+**Фильтрующие выражения:**
+
+- **type** — для фильтрации по значению соответствующего идентификатора (host, net или port);
+- **dir** — для фильтрации по конкретному направлению передачи к и / или от идентификатора (src и dst);
+- **proto** — для фильтрации по заданному сетевому протоколу (tcp, udp, icmp, ah и т.п.).
+
+
+**Примеры использования:**
+
+- Для вывода конкретного типа трафика – выражения tcp, udp, icmp:
+
+  **tcpdump icmp**
+
+- Трафик по протоколу (IPv4, IPv6):
+
+  **tcpdump ip6**
+
+- По диапазону портов – выражение portrange:
+
+  **tcpdump portrange 8000-9000**
+
+- Все пакеты определенного размера —выражения less или greater (либо соответствующие математические символы сравнения):
+
+  - Меньше 32 байт:
+
+    **tcpdump less 32**
+
+  - Больше 64 байт:
+
+    **tcpdump greater 64**
+
+- Отображение конкретных TCP-флагов:
+
+  **tcpdump 'tcp[13] & 32!=0'**
+
+- Выводпакетов ACKNOWLEDGE (ACK):
+
+  **tcpdump 'tcp[13] & 16!=0'**
+
+- Выводпакетов PUSH (PSH):
+
+  **tcpdump 'tcp[13] & 8!=0'**
+
+- Выводпакетов RESET (RST):
+
+  **tcpdump 'tcp[13] & 4!=0'**
+
+- Выводпакетов SYNCHRONIZE (SYN):
+
+  **tcpdump 'tcp[13] & 2!=0'**
+
+- Выводпакетов FINISH (FIN):
+
+  **tcpdump 'tcp[13] & 1!=0'**
+
+- Выводпакетов SYNCHRONIZE / ACKNOWLEDGE (SYNACK):
+
+  **tcpdump 'tcp[13]=18'**
+
+- Снятие флагов за счет tcpflags:
+
+  **tcpdump 'tcp[tcpflags] == tcp-**_ **flag** _ **'**
+
+- Поиск пакетов с нормальными комплектами RST и SYN:
+
+  **tcpdump 'tcp[13] = 6'**
+
+- Открытый текст HTTP и запрос:
+
+  **tcpdump 'tcp[32:4] = 0x47455420'**
+
+- SSH-соединения на конкретный порт (через баннер):
+
+  **tcpdump 'tcp[(tcp[12] \>\>2):4] = 0x5353482D'**
+
+- Пакеты С TTL \<10 (проблема или TRACEROUTE):
+
+  **tcpdump 'ip[8] \< 10'**
+
+- Пакеты с установкой EVIL BIT:
+
+  **tcpdump 'ip[6] & 128 != 0'**
+
+
+**Вывод результата.**
+
+Для сохранения результатов поиска пакетов в отдельный файл для дальнейшего анализа (файлы PCAP (PEE-cap)) – опция **-w** и имя файла.
+
+Чтобы открыть ранее сохраненные данные из файла PCAP —опция **-r** и имя файла.
+
+
+**Сравнение утилиты с procfs.**
+
+![image](https://github.com/Saeshnikov/Linux-monitoring-utility/assets/117933964/749a20c7-02d6-4cdb-93f1-3cc87edff1f5)
+
+Как видно из скриншота, утилита tcpdump обращается к файлам /proc только за файлом /proc/crypto/fips\_enabled. В файле /proc/crypto перечислены все установленные криптографические шифры, используемые ядром Linux, включая дополнительные сведения по каждому из них.
+
+
+**Актуальность утилиты.**
+
+Утилита tcpdump - удобный инструмент для решения проблем сетевого подключения. Несмотря на огромный вывод, который выдает утилита, ей удобно пользоваться благодаря множеству фильтров и выражений, которые можно комбинировать.
+
+Также, интерфейс командной строки tcpdump обеспечивает большую гибкость для сбора и анализа сетевого трафика. Её можно использовать в сочетании с графическим инструментом для понимания более сложных потоков – Wireshark: Wireshark может считывать файлы .pcap, захваченные tcpdump.
