+++ prototype/go.mod
@@ -1,0 +1,3 @@
+module linux-monitoring-utility
+
+go 1.22.0
