--- scripts/README.MD
@@ -1,1 +1,0 @@
-Скрипты для сборки, установки, анализа и прочих операций над проектом. Они позволяют оставить основной Makefile небольшим и простым.
